requirements.txt
requests 
from_root
boto3
python-dotenv
flask
flask-cors
ultralytics
taskflowai
-e .
